tools/requirements.txt
groq>=0.11.0
requests>=2.31.0
python-dotenv>=1.0.0
pyTelegramBotAPI>=4.18.0
